# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_parse_resume_api_v1_resumes_parse_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "PDF, DOCX, or UTF-8 text resume",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_parse_resume_api_v1_resumes_parse_post",
        "type": "object"
      },
...
      "DatePosted": {
        "description": "Portable date filters supported by job discovery.",
        "enum": [
          "all",
          "today",
          "3days",
          "week",
          "month"
        ],
        "title": "DatePosted",
        "type": "string"
      },
      "EmploymentType": {
        "description": "Supported employment types.",
        "enum": [
          "Full Time",
          "Part Time",
          "Internship",
          "Contract",
          "Remote"
        ],
        "title": "EmploymentType",
        "type": "string"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobResponse": {
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "employment_type": {
            "$ref": "#/components/schemas/EmploymentType"
          },
          "experience_level": {
            "$ref": "#/components/schemas/ExperienceLevel"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "required_skills": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SkillResponse"
            },
            "title": "Required Skills",
            "type": "array"
          },
          "source": {
            "$ref": "#/components/schemas/JobSource"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "company",
          "location",
          "description",
          "experience_level",
          "employment_type",
          "source",
          "url"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobSearchRequest": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "date_posted": {
            "$ref": "#/components/schemas/DatePosted",
            "default": "all"
          },
          "employment_types": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EmploymentType"
            },
            "title": "Employment Types",
            "type": "array"
          },
          "location": {
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "remote_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Only"
          },
          "role": {
            "minLength": 1,
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "location"
        ],
        "title": "JobSearchRequest",
        "type": "object"
      },
      "JobSearchResponse": {
        "properties": {
          "jobs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Jobs",
            "type": "array"
          },
          "provider_failures": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProviderFailureResponse"
            },
            "title": "Provider Failures",
            "type": "array"
          },
          "providers_attempted": {
            "title": "Providers Attempted",
            "type": "integer"
          },
          "providers_succeeded": {
            "title": "Providers Succeeded",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/JobSearchStatus"
          }
        },
        "required": [
          "status",
          "providers_attempted",
          "providers_succeeded"
        ],
        "title": "JobSearchResponse",
        "type": "object"
      },
      "JobSearchStatus": {
        "enum": [
          "complete",
          "partial",
          "failed"
        ],
        "title": "JobSearchStatus",
        "type": "string"
      },
      "JobSearchTaskCreatedResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BackgroundTaskStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "access_token",
          "status"
        ],
        "title": "JobSearchTaskCreatedResponse",
        "type": "object"
      },
      "JobSearchTaskResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "cancellation_requested": {
            "default": false,
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSearchResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/BackgroundTaskStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "attempt_count",
          "max_attempts",
          "created_at",
          "updated_at"
        ],
        "title": "JobSearchTaskResponse",
        "type": "object"
      },
...
      "ParsedResumeResponse": {
        "description": "One-time parsing response that returns user-submitted source text.",
        "properties": {
          "raw_text": {
            "title": "Raw Text",
            "type": "string"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeResponse"
          }
        },
        "required": [
          "resume",
          "raw_text"
        ],
        "title": "ParsedResumeResponse",
        "type": "object"
      },
      "ProviderFailureResponse": {
        "properties": {
          "code": {
            "default": "provider_failed",
            "title": "Code",
            "type": "string"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          }
        },
        "required": [
          "provider_name"
        ],
        "title": "ProviderFailureResponse",
        "type": "object"
      },
...
      "ResumeResponse": {
        "description": "Structured resume data safe for ordinary API responses.",
        "properties": {
          "achievements": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Achievements",
            "type": "array"
          },
          "certifications": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Certifications",
            "type": "array"
          },
          "education": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Education",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "experience": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Experience",
            "type": "array"
          },
          "github": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "linkedin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "projects": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Projects",
            "type": "array"
          },
          "skills": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SkillResponse"
            },
            "title": "Skills",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "ResumeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Typed HTTP API for Solara Hire.",
    "title": "Solara Hire",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs/search": {
      "post": {
        "operationId": "search_jobs_api_v1_jobs_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search configured job providers",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/search-tasks": {
      "post": {
        "operationId": "create_search_task_api_v1_jobs_search_tasks_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchTaskCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {},
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create an asynchronous job-discovery task",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/search-tasks/{task_id}": {
      "delete": {
        "operationId": "cancel_search_task_api_v1_jobs_search_tasks__task_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Task-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 20,
              "title": "X-Task-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {},
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request cancellation of an asynchronous job search",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_search_task_api_v1_jobs_search_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Task-Token",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 20,
              "title": "X-Task-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {},
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Poll an asynchronous job-discovery task",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a job returned by this API process",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/resumes/current": {
      "get": {
        "operationId": "get_current_resume_api_v1_resumes_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get the signed-in user's current resume profile",
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/v1/resumes/parse": {
      "post": {
        "operationId": "parse_resume_api_v1_resumes_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_api_v1_resumes_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParsedResumeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {},
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Parse a resume",
        "tags": [
          "resumes"
        ]
      }
    },
...
  }
}